requirements.txt
# BlackCell Security Toolkit Requirements

# TUI Framework
textual>=0.44.0
rich>=13.7.0
click>=8.1.7

# Network and Security Tools
requests>=2.31.0
urllib3>=2.1.0
python-nmap>=0.7.1
scapy>=2.5.0
paramiko>=3.3.1
cryptography>=41.0.0

# Web Security
beautifulsoup4>=4.12.2
selenium>=4.15.0
httpx>=0.25.0

# Fuzzing and Payloads
Wfuzz>=3.1.0
pyfiglet>=1.0.2

# System and File Operations
psutil>=5.9.6
pathlib2>=2.3.7
watchdog>=3.0.0

# Data Processing
pyyaml>=6.0.1
json5>=0.9.14
toml>=0.10.2

# Logging and Configuration
colorlog>=6.7.0
configparser>=6.0.0

# Testing Framework
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development Tools
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0